* commit '5bcf5942933e759ba888fdee3046e8ebf3bb0368':
  build error fix

+++ pkg/util/huawei/storagecache.go
@@ -180,7 +180,7 @@ func (self *SStoragecache) uploadImage(ctx context.Context, userCred mcclient.To
 	}
 
 	minDiskMB, _ := _image.Int("min_disk")
-	minDiskGB := int(math.Ceil(float64(minDiskMB) / 1024))
+	minDiskGB := int64(math.Ceil(float64(minDiskMB) / 1024))
 	// 在使用OBS桶的外部镜像文件制作镜像时生效且为必选字段。取值为40～1024GB。
 	if minDiskGB < 40 {
 		minDiskGB = 40
